% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yAdjustment.R
\name{calc_yAdjustment}
\alias{calc_yAdjustment}
\title{Calculate y axis adjustment (two comparisons only)}
\usage{
calc_yAdjustment(values_df, values_col, facet_var_value, constant)
}
\arguments{
\item{values_df}{'Melted' dataframe (long format) containing the actual values that will be plotted.}

\item{values_col}{Name of the column that contains the plotted values.}

\item{facet_var_value}{Name of the facet variable (column) in the values dataframe.}

\item{constant}{Constant value to be added for a standard distance between the brackets. Adjust according to expression values and number of statistical comparisons.}
}
\value{
The same values_df with an extra column called 'yadjustment' with y axis values.
}
\description{
Calculates the y axis adjustment positions for placement invisible points to improve plotting space to allow the placement of brackets.
}
\details{
Obs: The groups MUST come in the order that you want them to appear in the plot.
}
