% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yAdjustment.R
\name{calc_yAdjustment2}
\alias{calc_yAdjustment2}
\title{Calculate y axis adjustment (>2 comparisons)}
\usage{
calc_yAdjustment2(
  annotation_df,
  values_df,
  values_col,
  facet_var_value,
  facet_var_annot,
  group_var_annot,
  base_var_annot,
  stats_col,
  constant
)
}
\arguments{
\item{annotation_df}{Annotation dataframe to be used for plotting stats and brackets.}

\item{values_df}{'Melted' dataframe (long format) containing the actual values that will be plotted.}

\item{values_col}{Name of the column that contains the plotted values. Default value: "Expression".}

\item{facet_var_value}{Name of the facet variable (column) in the values dataframe. Default value: "Gene".}

\item{facet_var_annot}{Name of the facet variable (column) in the annotation dataframe. Default value: "Gene".}

\item{group_var_annot}{Name of the grouping variable (column) in the annotation dataframe. Default value: "Group2".}

\item{base_var_annot}{Name of the base grouping variable (column) in the annotation dataframe. This is the group of the reference level. Default value: "Group1".}

\item{stats_col}{Name of the statistics variable (column) in the annotation dataframe. Default value: "padj".}

\item{constant}{Constant value to be added for a standard distance between the brackets. Adjust according to expression values and number of statistical comparisons.}
}
\value{
The same values_df with an extra column called 'yadjustment' with y axis values.
}
\description{
Calculates the y axis adjustment positions for placement invisible points to improve plotting space to allow the placement of brackets.
}
\details{
Obs: The groups MUST come in the order that you want them to appear in the plot.
}
