% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yPosition.R
\name{calc_yPosition}
\alias{calc_yPosition}
\title{Calculate y axis position (two comparisons only)}
\usage{
calc_yPosition(
  annotation_df,
  values_df,
  values_col,
  facet_var_value,
  facet_var_annot,
  group_var_annot,
  constant
)
}
\arguments{
\item{annotation_df}{Annotation dataframe to be used for plotting stats and brackets.}

\item{values_df}{'Melted' dataframe (long format) containing the actual values that will be plotted.}

\item{values_col}{Name of the column that contains the plotted values.}

\item{facet_var_value}{Name of the facet variable (column) in the values dataframe.}

\item{facet_var_annot}{Name of the facet variable (column) in the annotation dataframe.}

\item{group_var_annot}{Name of the grouping variable (column) in the annotation dataframe.}

\item{constant}{Constant value to be added for a standard distance between the brackets. Adjust according to expression values and number of statistical comparisons.}
}
\value{
The same annotation_df with an extra column called 'yposition' with y axis values.
}
\description{
Calculates the y axis position for placement of brackets when adding stats to faceted plots from a melted matrix.
}
\details{
Obs: The groups MUST come in the order that you want them to appear in the plot.
}
