% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSVA.R
\name{cleanSVA}
\alias{cleanSVA}
\title{Regress out batch/unwanted variation using SVA}
\usage{
cleanSVA(y, mod, svaobj)
}
\arguments{
\item{y}{The counts matrix from which to remove the effects.}

\item{mod}{Model matrix to be tested (full model of the known effects). Must be the same used at \code{\link[=svaseq]{svaseq()}}}

\item{svaobj}{The result from \code{\link[=svaseq]{svaseq()}}.}
}
\value{
'Clean' matrix with unwanted effects removed.
}
\description{
Remove batch/unwanted effects from count data.
}
\details{
Takes the result from Surrogate Variable Analysis (see \code{\link[=svaseq]{svaseq()}}) and use it to regress out
the unwanted effects from count data.
This method was used by Jaffe and coleagues in the following paper: \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-015-0808-5}{10.1186/s12859-015-0808-5}.
}
