% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCheck.R
\name{geneCheck}
\alias{geneCheck}
\title{Check DESeq2 results}
\usage{
geneCheck(results_list, genes_df)
}
\arguments{
\item{results_list}{A named list containing 'DESeqResults' objects.}

\item{genes_df}{A dataframe of genes of interest where the column with gene symbols is named 'hgnc_symbol'.}
}
\value{
The function check every gene of genes_df against each of the results and aggregates log2 fold changes and adjusted p-values into a single data.frame for each result of the list. Results are rounded to two digits.
}
\description{
Compares results from DESeq2 and a predefined dataframe with previously selected genes for analysis.
}
