% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneNamesForPlotting.R
\name{geneNamesForPlotting}
\alias{geneNamesForPlotting}
\title{Gene names for plotting}
\usage{
geneNamesForPlotting(genes_df)
}
\arguments{
\item{genes_df}{A dataframe containing the genes analysed. Requires 'name' column with common name and 'hgnc_symbol' with the official gene symbol.}
}
\value{
The same dataframe with an extra column (names_plot column) with names in the format: SYMBOL(Alias).
}
\description{
Create labels for plotting to improve understanding of data by allowing the readers to identify genes by their common aliases.
}
