% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{get_lower_tri}
\alias{get_lower_tri}
\title{Get lower triangle}
\usage{
get_lower_tri(x)
}
\arguments{
\item{x}{The original matrix.}
}
\value{
The lower triangle of the original matrix.
}
\description{
Extracts the 'lower triangle' of a matrix.
}
