% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{Get upper triangle}
\usage{
get_upper_tri(x)
}
\arguments{
\item{x}{The original matrix.}
}
\value{
The upper triangle of the original matrix.
}
\description{
Extracts the 'upper triangle' of a matrix.
}
