% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in operator}
\usage{
... \%notin\% NA
}
\value{
True/False.
}
\description{
The reverse of the `%in%` in operator. Tells if a value is not present in another set o values.
}
