% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{is.numeric0}
\alias{is.numeric0}
\title{Check numeric(0)}
\usage{
is.numeric0(x)
}
\arguments{
\item{x}{The value/array to check.}
}
\value{
True or False.
}
\description{
Checks if something returns numeric(0). Posted by 'Andrii' user at https://stackoverflow.com/questions/10664662/how-to-test-when-condition-returns-numeric0-in-r
}
