% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterDESeq.R
\name{iterDESeq}
\alias{iterDESeq}
\title{DESeq with maxit argument}
\usage{
iterDESeq(obj, maxit = 500)
}
\arguments{
\item{obj}{The DESeq object on which to execute DESeq2()}

\item{maxit}{Number of maximum iterations to ther nbinomWaldTest test in DESeq2. Default is 500.}
}
\value{
The same DESeq object after running all the functions above.
}
\description{
Runs DESeq2 with the different maximum iterations value.
}
\details{
This function executes the same steps executed by the `DESeq2()` wrapper function, allowing you to change the maximum iterations value. These steps are:

`dds <- estimateSizeFactors(dds)`
`dds <- estimateDispersions(dds)`
`dds <- nbinomWaldTest(dds, maxit = n)`

Where `n` will be the chosen maximum iterations number in this function's `maxit` parameter.
}
