% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterDESeq2.R
\name{iterDESeq2}
\alias{iterDESeq2}
\title{DESeq2 with maxiter argument}
\usage{
iterDESeq2(obj, maxit = 500)
}
\arguments{
\item{obj}{The DESeq object on which to execute DESeq2()}

\item{maxit}{Number of maximum iterations to ther nbinomWaldTest test in DESeq2. Default is 500.}
}
\value{
DESeq2 object after running all the functions.
}
\description{
Runs DESeq2 with the different maximum iterations value. To do that, it runs separately the following functions:
}
\details{
dds <- estimateSizeFactors(dds)
dds <- estimateDispersions(dds)
dds <- nbinomWaldTest(dds, maxit = n)

Where n will be the chosen maximum iterations number.
}
