% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{matrixWithRowNames}
\alias{matrixWithRowNames}
\title{DF to Matrix with rownames.}
\usage{
matrixWithRowNames(x)
}
\arguments{
\item{x}{The dataframe to be converted.}
}
\value{
A matrix with the same rownames as the input dataframe.
}
\description{
Converts a dataframe to a matrix and keeps the rownames.
}
