% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{minMaxNorm}
\alias{minMaxNorm}
\title{Min/max normalization}
\usage{
minMaxNorm(x, min = NULL, max = NULL)
}
\arguments{
\item{x}{The vector of values to normalize according to the minimum and maximum values of the same vector.}

\item{min}{Minimum value in set of values. If min = NULL then min(x) is applied.}

\item{max}{Maximum value in set of values. If max = NULL then max(x) is applied.}
}
\value{
Normalized values in the vector.
}
\description{
Returns min/max normalized values.
}
