% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopDEGs.R
\name{plotTopDEGs}
\alias{plotTopDEGs}
\title{Plot top differentially expressed genes}
\usage{
plotTopDEGs(
  results,
  counts_matrix,
  n_genes = 20,
  which = "abs",
  logFC_col,
  pval_col,
  groups,
  pdata,
  sample_col = sample,
  group_col = group,
  split_col = TRUE
)
}
\arguments{
\item{results}{A data.frame or list of data.frames containing differential expression results.}

\item{counts_matrix}{Normalized and log transformed matrix of counts for plotting. Should be already centered.}

\item{n_genes}{Number of top DEGs to select. Default = \code{20}.}

\item{which}{A string indicating which direction to choose, \code{'up'} for overrepresented genes, \code{'down'} for downrepresented genes and \code{'abs'} for absolute values (both ways). Default: \code{'abs'}.}

\item{logFC_col}{The name of the column where the fold changes values are stored. Not a string.}

\item{pval_col}{Name of the column where the p-values (adjusted) are stored. Not a string.}

\item{groups}{A \code{data.frame} with results names as rownames (the same as names of results in list), and two columns in this order: one indicating the baseline group and one for the compared group. These must be the sames as in \code{pdata} argument \code{data.frame}.}

\item{pdata}{A \code{data.frame} containing \code{sample} and \code{groups} columns.}

\item{sample_col}{Name of column with sample information. Not a string.}

\item{group_col}{Name of column with group information. Not a string.}

\item{split_col}{\code{TRUE} or \code{FALSE}. Specifies wether to split the columns of the heatmaps based on sample information.}
}
\value{
A Heatmap or List of Heatmaps as per the \link{ComplexHeatmap} package.
}
\description{
Takes a list o results from differential expression pipelines and plot heatmaps of the top \code{n} differentially expressed
genes (DEGs) based on fold change value and plots them using \link{ComplexHeatmap} into a single heatmap or list of heatmaps.
The gene list can be calculated using positive fold changes, negative fold changes or absolute values.
}
