% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot preliminary Violin plots of Gene Expression data.}
\usage{
plotViolin(
  counts_matrix,
  results_df,
  genes_df,
  results_from = "deseq2",
  pheno_data,
  sample_col = "sample",
  control_group,
  control_color = "#33C4FF",
  test_group,
  test_color = "#FD4646",
  group_var,
  plot_order = NULL,
  n_rows = NULL,
  n_cols = NULL,
  alpha = 0.05
)
}
\arguments{
\item{counts_matrix}{A count matrix with official gene symbols as rownames.}

\item{results_df}{A results dataframe to obtain statists results from.}

\item{genes_df}{A dataframe containing list of genes to be analysed. Must contain at least one column named 'hgnc_symbol' with official gene symbols. The 'name' column is optional and is used to plot official names along with their aliases in a SYMBOL(Alias) format.}

\item{results_from}{Where the DGE results were calculated. Currently, the possible values are 'deseq2' (default) or 'limma'.}

\item{pheno_data}{Dataframe containin phenotypic data about the samples analysed.}

\item{control_group}{Name of the baseline group for comparison in pheno_data.}

\item{control_color}{Color for plotting the baseline group.}

\item{test_group}{Name of the tested group for comparison in pheno_data.}

\item{test_color}{Color for plotting the tested group.}

\item{group_var}{Name of the group variable in pheno_data.}

\item{n_rows}{If specified, defines the number of rows of plots in the final plotting area.}

\item{n_cols}{If specified, defines the number of columns of plots in the final plotting area.}

\item{alpha}{Significance threshold (Default = 0.05).}

\item{plots_order}{If specified, defines the order in which the genes must be plotted.}
}
\value{
A list object with three objects:

• data = The genes_df dataframe with new gene information.

• plot_annotation = The annotation_df used for plotting.

• violin_plot = The violin plots generated from the data as faceted ggplot2 plots.
}
\description{
This function takes a list a genes and extracts values from a count matrix to produce Violin plots of the selected genes to be used as a preliminary analysis.
Currently supports only comparisons between two groups.
}
