% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueCodes.R
\name{pvalueCodes}
\alias{pvalueCodes}
\title{Convert p-values to * codes.}
\usage{
pvalueCodes(annotation_df, pvalue_col)
}
\arguments{
\item{annotation_df}{Annotation dataframe constaining stats values.}

\item{pvalue_col}{Name of the column that contains the p-values.}
}
\value{
The same annotation_df with an extra column called 'pcode' with coded p-values.
}
\description{
Converts p-values to standard * codes.
}
\details{
p < 0.05 = *
p < 0.01 = **
p < 0.001 = ***
p < 0.0001 = ****
}
