% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yPosition.R
\name{calc_yPosition2}
\alias{calc_yPosition2}
\title{Calculate y axis position (>2 comparisons)}
\usage{
calc_yPosition2(
  annotation_df,
  values_df,
  values_col = "Expression",
  facet_var_value = "Gene",
  facet_var_annot = "Gene",
  group_var_annot = "Group2",
  base_var_annot = "Group1",
  stats_col = "padj",
  k,
  decrease = 0
)
}
\arguments{
\item{annotation_df}{Annotation dataframe to be used for plotting stats and brackets.}

\item{values_df}{'Melted' dataframe (long format) containing the actual values that will be plotted.}

\item{values_col}{Name of the column that contains the plotted values. Default value: "Expression".}

\item{facet_var_value}{Name of the facet variable (column) in the values dataframe. Default value: "Gene".}

\item{facet_var_annot}{Name of the facet variable (column) in the annotation dataframe. Default value: "Gene".}

\item{group_var_annot}{Name of the grouping variable (column) in the annotation dataframe. Default value: "Group2".}

\item{base_var_annot}{Name of the base grouping variable (column) in the annotation dataframe. This is the group of the reference level. Default value: "Group1".}

\item{stats_col}{Name of the statistics variable (column) in the annotation dataframe. Default value: "padj".}

\item{k}{Constant value to be added for a standard distance between the brackets. Adjust according to expression values and number of statistical comparisons.}

\item{decrease}{Integer number to subtract from total number of comparisons. This is for use in case you are plotting stats for less than the maximum amount of comparisons between the groups. Default is \code{0}.}
}
\value{
The same annotation_df with an extra column called 'yposition' with y axis values.
}
\description{
Calculates the y axis position for placement of brackets when adding stats to faceted plots from a melted matrix.
}
\details{
Obs: The groups MUST come in the order that you want them to appear in the plot.
}
