% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterDESeq.R
\name{iterDESeq}
\alias{iterDESeq}
\title{DESeq with maxit argument}
\usage{
iterDESeq(obj, maxit = 500)
}
\arguments{
\item{obj}{The DESeq object on which to execute DESeq2()}

\item{maxit}{Number of maximum iterations to ther nbinomWaldTest test in DESeq2. Default is 500.}
}
\value{
The same DESeq object after running all the functions above.
}
\description{
Runs DESeq2 with the different maximum iterations value.
}
\details{
This function executes the same steps executed by the \code{DESeq2()} wrapper function, allowing you to change the maximum iterations value. These steps are:

\code{dds <- estimateSizeFactors(dds)}
\code{dds <- estimateDispersions(dds)}
\code{dds <- nbinomWaldTest(dds, maxit = n)}

Where \code{n} will be the chosen maximum iterations number in this function's \code{maxit} parameter.
}
